Rename view.util to view.viewmodel
