+++ lyrebird/src/main/java/moe/lyrebird/view/viewmodel/Clipping.java
@@ -16,7 +16,7 @@
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package moe.lyrebird.view.util;
+package moe.lyrebird.view.viewmodel;
 
 import javafx.scene.shape.Circle;
 import javafx.scene.shape.Rectangle;
